# services/api/server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: MSPMR
  description: MSPMR Media Share API
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
definitions:
  VideoBody:
    type: object
    title: video
    required:
      - name
    properties:
      name:
        type: string
        video: no_stress
paths:
...
  /videos/{id}:
    get:
      tags:
        - Videos
      description: Get an video
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the video to retrieve
          type: string
      responses:
        200:
          description: Return the video with the specified id
        404:
          description: Video not found
    patch:
      tags:
        - Videos
      description: Update an video
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the video to update
          type: string
        - name: video
          in: body
          description: video attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/VideoBody"
      responses:
        200:
          description: Return updated videos
        404:
          description: Video not found
    delete:
      tags:
        - Videos
      description: Delete an video
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the video to delete
          type: string
      responses:
        204:
          description: Delete the video with the specified id
        404:
          description: Video not found
...